Alert box!
feedback from Max Airey, when max record reached, should not have alert box that stops the timer. Instead, a pop up should be used while not stopping the counting so that user can know what to do without losing progress. Also, updated the displaying of clock and timer component that when one is toggle, the other will now hide instead of fully refresh. Before if you toggle timer and start it, and then you toggle clock, all your progress and record in timer page will be wiped out. This is not wanted and therefore an update maded to allow free toggle between two component while saving clients data and options.